categories working

+++ src/main/java/com/codeup/adlister/dao/Users.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface Users {
     User findByUsername(String username);
-    Long insert(User user);
+    int insert(User user);
 }
